niyam-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
# supabase==1.0.3 # Optional if using mock
# pydantic==2.5.0 # Installed with fastapi
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# pytesseract==0.3.10
# pdf2image==1.16.3
# Pillow==10.1.0
# redis==5.0.1
# celery==5.3.4
# pandas==2.1.4
# numpy==1.26.2
python-dateutil==2.8.2
httpx==0.25.1
email-validator==2.1.0
